Extract static Message format functions to new MessageFormatter class
