+++ Pulsar4X/Pulsar4X.ECSLib/Messages/MessagePump.cs
@@ -14,12 +14,6 @@ namespace Pulsar4X.ECSLib
     {
         private readonly ConcurrentQueue<string> _incomingMessages = new ConcurrentQueue<string>();
         private readonly ConcurrentQueue<string> _outgoingMessages = new ConcurrentQueue<string>();
-
-        private static readonly Regex _guidListRegex = new Regex("^(?:([\\w]+),?)*;");
-        private static readonly Regex _guidRegex = new Regex("^(?:([\\w]+);)");
-        private static readonly Regex _messageTypeRegex = new Regex("^(\\d+);");
-        private static readonly Regex _authTokenRegex = new Regex("^(\\w+\n\\w+)\n");
-
         #region Queue Management
 
         /// <summary>
@@ -36,7 +30,7 @@ namespace Pulsar4X.ECSLib
         [PublicAPI]
         public void EnqueueMessage(IncomingMessageType messageType, AuthenticationToken authToken, string message)
         {
-            message = GetMessageHeader(messageType, authToken) + message; 
+            message = MessageFormatter.GetMessageHeader(messageType, authToken) + message; 
             EnqueueMessage(message);
         }
 
@@ -60,9 +54,11 @@ namespace Pulsar4X.ECSLib
                 throw new InvalidEnumArgumentException(nameof(messageType), (int)messageType, typeof(OutgoingMessageType));
             }
 
-            message = GetOutgoingMessageHeader(messageType) + message; 
+            message = MessageFormatter.GetOutgoingMessageHeader(messageType) + message; 
             _outgoingMessages.Enqueue(message);
         }
+
+        internal void EnqueueOutgoingMessage(string message) => _outgoingMessages.Enqueue(message);
 
         /// <summary>
         /// Attempts to peek a message from the ECSLib's Outgoing Message Queue. Does not remove the message from the queue.
@@ -76,163 +72,5 @@ namespace Pulsar4X.ECSLib
         [PublicAPI]
         public bool TryDequeueOutgoingMessage(out string message) => _outgoingMessages.TryDequeue(out message);
         #endregion Queue Management
-
-        #region Message Deconstruction
-
-        /// <summary>
-        /// Retrieves a list of of Guid's from the message.
-        /// Strips the detected guid list from the message for further processing.
-        /// </summary>
-        [PublicAPI]
-        public static bool TryGetGuidList(ref string message, out List<Guid> guidList)
-        {
-            guidList = new List<Guid>();
-
-            // Detect if the passed message has a CSV GuidList at the beginning
-            Match match = _guidListRegex.Match(message);
-            if (!match.Success)
-            {
-                return false;
-            }
-
-            foreach (Capture capture in match.Groups[1].Captures)
-            {
-                guidList.Add(Guid.Parse(capture.Value));
-            }
-
-            // Remove the captured guidList from the message.
-            message = message.Substring(match.Length);
-            return true;
-        }
-
-        /// <summary>
-        /// Retrieves a Guid from the message.
-        /// Strips the detected Guid from the message for further processing.
-        /// </summary>
-        [PublicAPI]
-        public static bool TryGetGuid(ref string message, out Guid guid)
-        {
-            guid = Guid.Empty;
-            Match match = _guidRegex.Match(message);
-            if (!match.Success)
-            {
-                return false;
-            }
-
-            guid = Guid.Parse(match.Groups[1].Captures[0].Value);
-
-            message = message.Substring(match.Length);
-            return true;
-        }
-
-        /// <summary>
-        /// Converts a list of Guids to a string formatted to be read by TryGetGuidList
-        /// </summary>
-        [PublicAPI]
-        public static string GuidListToString(IEnumerable<Guid> guidList)
-        {
-            var builder = new StringBuilder();
-            foreach (Guid guid in guidList)
-            {
-                builder.Append($"{guid:N},");
-            }
-            builder.Replace(",", ";", builder.Length - 1, 1);
-            return builder.ToString();
-        }
-
-        /// <summary>
-        /// Retrieves the OutgoingMessageType from the provided message.
-        /// Strips the OutgoingMessageType from the message for further processing.
-        /// </summary>
-        [PublicAPI]
-        public static bool TryGetOutgoingMessageType(ref string message, out OutgoingMessageType messageType)
-        {
-            messageType = OutgoingMessageType.Invalid;
-            Match match = _messageTypeRegex.Match(message);
-            if (!match.Success || !Enum.TryParse(match.Groups[1].Captures[0].Value, out messageType))
-            {
-                return false;
-            }
-
-            // Strip the OutgoingMessageType from the message.
-            message = message.Substring(match.Length);
-            return true;
-        }
-
-        private static bool TryGetIncomingMessageType(ref string message, out IncomingMessageType messageType)
-        {
-            messageType = IncomingMessageType.Invalid;
-            Match match = _messageTypeRegex.Match(message);
-            if (!match.Success || !Enum.TryParse(match.Groups[1].Captures[0].Value, out messageType))
-            {
-                return false;
-            }
-
-            // Strip the IncomingMessageType from the message.
-            message = message.Substring(match.Length);
-            return true;
-        }
-
-        private static bool TryGetAuthToken(ref string message, out AuthenticationToken authToken)
-        {
-            authToken = null;
-            Match match = _authTokenRegex.Match(message);
-            if (!match.Success)
-            {
-                return false;
-            }
-
-            authToken = new AuthenticationToken(match.Groups[1].Captures[0].Value);
-            message = message.Substring(match.Length);
-            return true;
-        }
-
-        /// <summary>
-        /// Retrieves the IncomingMessageType and AuthenticationToken from the provided message.
-        /// Strips the header from the message for additional processing.
-        /// </summary>
-        internal static bool TryDeconstructHeader(ref string message, out IncomingMessageType messageType, out AuthenticationToken authToken)
-        {
-            string originalMessage = message;
-            authToken = null;
-
-            if (TryGetIncomingMessageType(ref message, out messageType) && TryGetAuthToken(ref message, out authToken))
-            {
-                return true;
-            }
-
-            // Invalid Header. Reconstruct original message.
-            message = originalMessage;
-            return false;
-        }
-
-        #endregion Message Deconstruction
-
-        #region Message Construction
-
-        /// <summary>
-        /// Returns a validly formatted header with the provided messageType and authToken.
-        /// </summary>
-        /// <remarks>
-        /// Use this function to construct your message headers so the UI doesn't have to worry about how the ECSLib expects the header to be formatted.
-        /// </remarks>
-        [PublicAPI]
-        public static string GetMessageHeader(IncomingMessageType messageType, [NotNull] AuthenticationToken authToken)
-        {
-            if (authToken == null)
-            {
-                throw new ArgumentNullException(nameof(authToken));
-            }
-            if (!Enum.IsDefined(typeof(IncomingMessageType), messageType))
-            {
-                throw new InvalidEnumArgumentException(nameof(messageType), (int)messageType, typeof(IncomingMessageType));
-            }
-
-            return $"{(int)messageType};{authToken}\n";
-        }
-
-        private static string GetOutgoingMessageHeader(OutgoingMessageType messageType) => $"{(int)messageType};";
-        #endregion Message Construction
-
     }
 }
