+++ Pulsar4X/Pulsar4X.Tests/MessageTests.cs
@@ -35,7 +35,7 @@ namespace Pulsar4X.Tests
             Assert.AreEqual(outString, incomingMessage);
             Assert.IsTrue(_messagePump.TryDequeueOutgoingMessage(out outString));
             OutgoingMessageType outMessageTypeTest;
-            Assert.IsTrue(MessagePump.TryGetOutgoingMessageType(ref outString, out outMessageTypeTest));
+            Assert.IsTrue(MessageFormatter.TryGetOutgoingMessageType(ref outString, out outMessageTypeTest));
             Assert.AreEqual(outMessageTypeTest, OutgoingMessageType.Invalid);
             Assert.AreEqual(outString, outgoingMessage);
 
@@ -53,7 +53,7 @@ namespace Pulsar4X.Tests
             IncomingMessageType incomingMessageTypeTest;
             AuthenticationToken authTokenTest;
 
-            Assert.IsTrue(MessagePump.TryDeconstructHeader(ref outString, out incomingMessageTypeTest, out authTokenTest));
+            Assert.IsTrue(MessageFormatter.TryDeconstructHeader(ref outString, out incomingMessageTypeTest, out authTokenTest));
             Assert.AreEqual(authToken, authTokenTest);
             Assert.AreEqual(IncomingMessageType.Echo, incomingMessageTypeTest);
             Assert.AreEqual(message, outString);
@@ -68,12 +68,12 @@ namespace Pulsar4X.Tests
             // Retrieve a header for a message.
             var messageType = IncomingMessageType.Echo;
             var authToken = new AuthenticationToken(Guid.NewGuid(), "hunter2");
-            string message = MessagePump.GetMessageHeader(messageType, authToken);
+            string message = MessageFormatter.GetMessageHeader(messageType, authToken);
 
             // Decode the header we retrieved above, ensure equality.
             IncomingMessageType testMessageType;
             AuthenticationToken testAuthToken;
-            MessagePump.TryDeconstructHeader(ref message, out testMessageType, out testAuthToken);
+            MessageFormatter.TryDeconstructHeader(ref message, out testMessageType, out testAuthToken);
             Assert.IsEmpty(message);
             Assert.AreEqual(messageType, testMessageType);
             Assert.AreEqual(authToken, testAuthToken);
@@ -83,7 +83,7 @@ namespace Pulsar4X.Tests
             OutgoingMessageType outMessageTypeTest;
 
             Assert.True(_messagePump.TryPeekOutgoingMessage(out message));
-            Assert.IsTrue(MessagePump.TryGetOutgoingMessageType(ref message, out outMessageTypeTest));
+            Assert.IsTrue(MessageFormatter.TryGetOutgoingMessageType(ref message, out outMessageTypeTest));
             Assert.AreEqual(message, "EchoTest");
             Assert.AreEqual(OutgoingMessageType.Echo, outMessageTypeTest);
         }
@@ -97,7 +97,7 @@ namespace Pulsar4X.Tests
             var messageType = (IncomingMessageType)(-1);
             var authToken = new AuthenticationToken(Guid.NewGuid(), "hunter2");
             string message;
-            Assert.Throws<InvalidEnumArgumentException>(() => message = MessagePump.GetMessageHeader(messageType, authToken));
+            Assert.Throws<InvalidEnumArgumentException>(() => message = MessageFormatter.GetMessageHeader(messageType, authToken));
 
             // Attempt to manually queue a message with a invalid IncomingMessageType.
             // NOTE: This is a valid situation the MessageDispatcher will have to deal with if VM sends a malformed message.
@@ -107,17 +107,17 @@ namespace Pulsar4X.Tests
 
             IncomingMessageType messageTypeTest;
             AuthenticationToken authTokenTest;
-            Assert.IsFalse(MessagePump.TryDeconstructHeader(ref message, out messageTypeTest, out authTokenTest));
+            Assert.IsFalse(MessageFormatter.TryDeconstructHeader(ref message, out messageTypeTest, out authTokenTest));
 
             // Attempt to get a header for a null authToken.
-            Assert.Throws<ArgumentNullException>(() => invalidMsg = MessagePump.GetMessageHeader(IncomingMessageType.Echo, null));
+            Assert.Throws<ArgumentNullException>(() => invalidMsg = MessageFormatter.GetMessageHeader(IncomingMessageType.Echo, null));
 
             // Attempt to manually queue a message with an invalid authToken
             // NOTE: This is a valid situation the MessageDispatcher will have to deal with if VM sends a malformed message.
             invalidMsg = "1;\n";
             _messagePump.EnqueueMessage(invalidMsg);
             Assert.IsTrue(_messagePump.TryDequeueIncomingMessage(out message));
-            Assert.IsFalse(MessagePump.TryDeconstructHeader(ref message, out messageTypeTest, out authTokenTest));
+            Assert.IsFalse(MessageFormatter.TryDeconstructHeader(ref message, out messageTypeTest, out authTokenTest));
 
             // Attempt to queue a message with an invalid OutgoingMessageType.
             Assert.Throws<InvalidEnumArgumentException>(() => _messagePump.EnqueueOutgoingMessage((OutgoingMessageType)(-1), "Invalid MessageType"));
@@ -127,7 +127,7 @@ namespace Pulsar4X.Tests
             // NOTE: This is a valid situation the VM will have to deal with if the ECSLib sends a malformed message.
             invalidMsg = "-1;";
             OutgoingMessageType outMessageTypeTest;
-            Assert.False(MessagePump.TryGetOutgoingMessageType(ref invalidMsg, out outMessageTypeTest));
+            Assert.False(MessageFormatter.TryGetOutgoingMessageType(ref invalidMsg, out outMessageTypeTest));
         }
 
         [Test]
@@ -141,11 +141,11 @@ namespace Pulsar4X.Tests
             }
 
             // Convert GuidList to string.
-            string guidListString = MessagePump.GuidListToString(guidList);
+            string guidListString = MessageFormatter.GuidListToString(guidList);
 
             // Convert string to new GuidList
             List<Guid> retrievedGuids;
-            Assert.IsTrue(MessagePump.TryGetGuidList(ref guidListString, out retrievedGuids));
+            Assert.IsTrue(MessageFormatter.TryGetGuidList(ref guidListString, out retrievedGuids));
 
             // Verify we got the same Guids back
             for (int i = 0; i < 10; i++)
@@ -165,7 +165,7 @@ namespace Pulsar4X.Tests
             }
             guidListString.Substring(0, guidListString.Length - 1);
             guidListString += ';';
-            Assert.IsFalse(MessagePump.TryGetGuidList(ref guidListString, out retrievedGuids));
+            Assert.IsFalse(MessageFormatter.TryGetGuidList(ref guidListString, out retrievedGuids));
         }
 
         [Test]
@@ -178,7 +178,7 @@ namespace Pulsar4X.Tests
 
             // Parse the Guid from the string.
             Guid parsedGuid;
-            Assert.IsTrue(MessagePump.TryGetGuid(ref guidString, out parsedGuid));
+            Assert.IsTrue(MessageFormatter.TryGetGuid(ref guidString, out parsedGuid));
 
             // Verify got the same Guid back.
             Assert.AreEqual(guid, parsedGuid);
@@ -187,7 +187,7 @@ namespace Pulsar4X.Tests
 
             // Ensure TryGetGuid returns false with a bad format.
             guidString = $"{guid:B};";
-            Assert.IsFalse(MessagePump.TryGetGuid(ref guidString, out parsedGuid));
+            Assert.IsFalse(MessageFormatter.TryGetGuid(ref guidString, out parsedGuid));
         }
     }
 }
